Image Preprocessing/requirements.txt

---

## 📦 `requirements.txt`

```text
streamlit
opencv-python
numpy
Pillow
matplotlib

